requirements.txt
torch
transformers
flask
python-dotenv
requests
numpy
pandas
accelerate
bitsandbytes
sentencepiece